refactor(package): 흩어져 있던 데이터 계층의 모델들을 데이터 계층으로 이동
