--- app/src/main/java/woowacourse/shopping/data/model/product/ProductsDto.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.data.model.product
-
-class ProductsDto : ArrayList<ProductDto>()
