+++ app/src/main/java/woowacourse/shopping/data/order/model/FixedDiscountPoliciesResponse.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.model.order
+package woowacourse.shopping.data.order.model
 
 data class FixedDiscountPoliciesResponse(
     val fixedDiscountPolicies: List<FixedDiscountPolicyResponse>
