+++ app/src/main/java/woowacourse/shopping/data/product/RetrofitProductService.kt
@@ -5,7 +5,7 @@ import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
-import woowacourse.shopping.data.model.product.ProductsResponse
+import woowacourse.shopping.data.product.model.ProductsResponse
 
 interface RetrofitProductService {
