+++ app/src/main/java/woowacourse/shopping/data/product/model/ProductsDto.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.data.product.model
+
+class ProductsDto : ArrayList<ProductDto>()
