+++ app/src/main/java/woowacourse/shopping/model/mapper/CartProductMapper.kt
@@ -1,7 +1,6 @@
 package woowacourse.shopping.model.mapper
 
 import com.example.domain.CartProduct
-import woowacourse.shopping.model.CartProductResponse
 import woowacourse.shopping.model.CartProductState
 
 fun CartProduct.toUi(): CartProductState {
@@ -25,17 +24,5 @@ fun CartProductState.toDomain(): CartProduct {
         productPrice = productPrice,
         quantity = quantity,
         isPicked = isPicked
-    )
-}
-
-fun CartProductResponse.toDomain(): CartProduct {
-    return CartProduct(
-        id = id,
-        productId = product.id,
-        productImageUrl = product.imageUrl,
-        productName = product.name,
-        productPrice = product.price,
-        quantity = quantity,
-        isPicked = true
     )
 }
