+++ app/src/main/java/woowacourse/shopping/model/mapper/FixedDiscountPoliciesMapper.kt
@@ -1,7 +1,6 @@
 package woowacourse.shopping.model.mapper
 
 import com.example.domain.FixedDiscountPolicies
-import woowacourse.shopping.model.order.FixedDiscountPoliciesResponse
 import woowacourse.shopping.model.order.FixedDiscountPoliciesState
 
 fun FixedDiscountPolicies.toUi(): FixedDiscountPoliciesState {
@@ -11,12 +10,6 @@ fun FixedDiscountPolicies.toUi(): FixedDiscountPoliciesState {
 }
 
 fun FixedDiscountPoliciesState.toDomain(): FixedDiscountPolicies {
-    return FixedDiscountPolicies(
-        fixedDiscountPolicies = fixedDiscountPolicies.map { it.toDomain() }
-    )
-}
-
-fun FixedDiscountPoliciesResponse.toDomain(): FixedDiscountPolicies {
     return FixedDiscountPolicies(
         fixedDiscountPolicies = fixedDiscountPolicies.map { it.toDomain() }
     )
