+++ app/src/main/java/woowacourse/shopping/model/mapper/OrderMapper.kt
@@ -2,20 +2,9 @@ package woowacourse.shopping.model.mapper
 
 import com.example.domain.order.Order
 import com.example.domain.order.OrderProduct
-import com.example.domain.order.OrderSummary
 import woowacourse.shopping.model.ProductState
 import woowacourse.shopping.model.order.OrderProductState
 import woowacourse.shopping.model.order.OrderState
-import woowacourse.shopping.model.order.OrderSummaryResponse
-
-fun OrderSummaryResponse.toDomain(): OrderSummary {
-    return OrderSummary(
-        id = id,
-        finalPrice = finalPrice,
-        products = products,
-        orderDate = orderDate
-    )
-}
 
 fun Order.toUi(): OrderState {
     return OrderState(
